tp3

+++ src/controllers/carts.js
@@ -47,10 +47,19 @@ export const addProductCart = async (req= request, res= response) => {
 
 export const deleteProductsInCart = async (req= request, res= response) => {
     try {
+        const {_id} = req;
         const {cid,pid} = req.params;
+
+        const usuario = await UsersRepository.getUserById(_id);
+        if(!usuario) return res.status(400).json({ok: false, msg:'Usuario no existe'});
+
+        if(!(usuario.cart_id.toString() === cid)) return res.status(400).json({ok:false, msg: 'Carrito no valido'});
+        
+        const existeProducto = await ProductsRepository.getProductById(pid);
+        if(!existeProducto) return res.status(400).json({ok:false, msg: 'Producto no existe'});
+
         const carrito = await CartsRepository.deleteProductsInCart(cid,pid);
-        if(!carrito)
-            return res.status(404).json({msg: 'No se pudo realizar esa operacion'})
+
         return res.json({msg: ' Producto eliminado del carrito', carrito})
     } catch (error) {
         return res.status(500).json({msg:"Hablar con admin"})
